refactor: presigned 코드 정리
#8
