+++ src/main/java/net/detalk/api/service/FileService.java
@@ -3,8 +3,7 @@ package net.detalk.api.service;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import net.detalk.api.domain.AttachmentFile;
-import net.detalk.api.domain.UploadFileMetadata;
-import net.detalk.api.domain.FileWithPresigned;
+import net.detalk.api.domain.PreSignedData;
 import net.detalk.api.repository.AttachmentFileRepository;
 import net.detalk.api.support.TimeHolder;
 import net.detalk.api.support.UUIDGenerator;
@@ -22,11 +21,10 @@ public class FileService {
     private final TimeHolder timeHolder;
     private final UUIDGenerator uuidGenerator;
 
-    public FileWithPresigned createPreSignedUrl(Long uploaderId, UploadFileMetadata uploadFileMetadata) {
+    public PreSignedData createPreSignedUrl(Long uploaderId, String fileName, String fileType, String type) {
         UUID fileId = uuidGenerator.generateV7();
 
-        String baseKey = String.format("u/%s", fileId);
-        String fileName = uploadFileMetadata.fileName();
+        String path = String.format("%s/%s/%s", fileType.split("/")[0], type, fileId);
 
         // 파일 이름과 확장자 분리
         String extension = null;
@@ -46,8 +44,7 @@ public class FileService {
         }
 
         // Pre-signed URL 생성
-        String preSignedUrl = storageClient.createPreSignedUrl(baseKey);
-        String publicUrl = String.format("https://cdn.detalk.net/" + baseKey);
+        String preSignedUrl = storageClient.createPreSignedUrl(path);
 
         AttachmentFile attachmentFile = attachmentFileRepository.save(
             AttachmentFile.builder()
@@ -56,13 +53,13 @@ public class FileService {
                 .name(realFileName)
                 .extension(extension)
                 .createdAt(timeHolder.now())
-                .url(publicUrl)
+                .url(path)
                 .build()
         );
 
-        return FileWithPresigned.builder()
+        return PreSignedData.builder()
             .id(attachmentFile.getId())
-            .url(publicUrl)
+            .path(path)
             .preSignedUrl(preSignedUrl)
             .build();
     }
